+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 Changelog
 =========
 
+# UNRELEASED
+
+## Added
+- Include X-Request-ID if given in the headers.
+
 
 # 0.1.0 - 2017-09-02